backend/sgventas/requirements/base.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
black==23.11.0
certifi==2022.12.7
cfgv==3.4.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.1.6
django-cors-headers==3.13.0
django-crum==0.7.9
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-rest-swagger==2.2.0
django-silk==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==17.3.0
filelock==3.13.1
flake8==6.0.0
gprof2dot==2022.7.29
identify==2.5.31
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
nodeenv==1.8.0
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.2
Pillow==9.4.0
platformdirs==3.11.0
pluggy==1.0.0
pre-commit==3.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.24.6
